Closes #61

This branch adds a global registry that tracks callsites and
subscribers. When a new callsite is registered, all subscribers are
informed of it immediately; similarly, when a new subscriber is
registered, it is informed of (and allowed to filter) all callsites.
This is used to update the filtering logic for callsites, and also
allows an opportunity for subscribers to observe a callsite's metadata
in a function that's guaranteed to be called exactly once. 

Exactly-once registration can be useful as a time where field keys 
can be extracted from the metadata and stored, ensuring that the
more costly by-name key lookup isn't repeated.

I've added a new benchmark for creating a span wuith the same callsite
multiple times, as this is where this branch offers the largest
performance improvement. 

As the becnhmark results below show, performance in that case
(`span_repeatedly`) is about 1/4th faster after this change.

Benchmarks before:

```
     Running target/release/deps/no_subscriber-30ffe0f7fbecf1ec

running 4 tests
test bench_1_atomic_load              ... bench:           0 ns/iter (+/- 0)
test bench_costly_field_no_subscriber ... bench:          48 ns/iter (+/- 9)
test bench_no_span_no_subscriber      ... bench:           0 ns/iter (+/- 0)
test bench_span_no_subscriber         ... bench:          44 ns/iter (+/- 13)

test result: ok. 0 passed; 0 failed; 0 ignored; 4 measured; 0 filtered out

     Running target/release/deps/subscriber-ab6525f0578fb401

running 4 tests
test span_no_fields            ... bench:         103 ns/iter (+/- 10)
test span_repeatedly           ... bench:      21,876 ns/iter (+/- 9,254)
test span_with_fields          ... bench:         278 ns/iter (+/- 175)
test span_with_fields_add_data ... bench:       1,026 ns/iter (+/- 163)
```

Benchmarks after:
```
     Running target/release/deps/no_subscriber-0122e236a67454e8

running 4 tests
test bench_1_atomic_load              ... bench:           0 ns/iter (+/- 0)
test bench_costly_field_no_subscriber ... bench:          37 ns/iter (+/- 7)
test bench_no_span_no_subscriber      ... bench:           0 ns/iter (+/- 0)
test bench_span_no_subscriber         ... bench:          33 ns/iter (+/- 5)

test result: ok. 0 passed; 0 failed; 0 ignored; 4 measured; 0 filtered out

     Running target/release/deps/subscriber-28d74558d25f84dc

running 4 tests
test span_no_fields            ... bench:          72 ns/iter (+/- 12)
test span_repeatedly           ... bench:      15,395 ns/iter (+/- 1,092)
test span_with_fields          ... bench:         155 ns/iter (+/- 14)
test span_with_fields_add_data ... bench:         698 ns/iter (+/- 87)
```

Signed-off-by: Eliza Weisman <eliza@buoyant.io>