@davidbarsky 2018-11-09
Error on CI:

```
error[E0432]: unresolved import `tower_service::NewService`
 --> /home/travis/.cargo/git/checkouts/tower-h2-aaf68b3c8982de4c/4e1a50a/src/client/connect.rs:4:5
  |
4 | use tower_service::NewService;
  |     ^^^^^^^^^^^^^^^^^^^^^^^^^ no `NewService` in the root. Did you mean to use `Service`?
error[E0432]: unresolved import `tower_service::NewService`
 --> /home/travis/.cargo/git/checkouts/tower-h2-aaf68b3c8982de4c/4e1a50a/src/server/mod.rs:4:21
  |
4 | use tower_service::{NewService, Service};
  |                     ^^^^^^^^^^ no `NewService` in the root. Did you mean to use `Service`?
```