@hawkw 2018-11-09
@davidbarsky sorry about the CI breakage, that was due to an upstream dependency change. I _thought_ fd21e4bdff2c611fc46b2b7d7d39da70e5f3ba3e would be sufficient to fix it, but it wasn't --- 5b55938a4988dc8cc5972b0140678c7f35c2c7f1 should actually resolve the problem.

Ah, the perils of git dependencies!